split up the source into multiple files and clients
